src/data_analytics.py
```python
import pandas as pd
from src.ai_tools import generate_recommendations, automate_management

class DataAnalytics:
    def __init__(self, user_data, investment_data, property_data, reit_data):
        self.user_data = pd.DataFrame(user_data)
        self.investment_data = pd.DataFrame(investment_data)
        self.property_data = pd.DataFrame(property_data)
        self.reit_data = pd.DataFrame(reit_data)

    def analyze_user_behavior(self, user_id):
        user_investments = self.investment_data[self.investment_data['user_id'] == user_id]
        user_properties = self.property_data[self.property_data['property_id'].isin(user_investments['property_id'])]
        user_reits = self.reit_data[self.reit_data['reit_id'].isin(user_investments['reit_id'])]

        # Analyze user's investment preferences based on their past investments
        preferred_property_type = user_properties['property_type'].mode()[0]
        preferred_reit = user_reits['reit_name'].mode()[0]

        return {'preferred_property_type': preferred_property_type, 'preferred_reit': preferred_reit}

    def identify_areas_for_improvement(self):
        # Identify users who have not made an investment yet
        users_without_investments = self.user_data[~self.user_data['user_id'].isin(self.investment_data['user_id'])]

        # Identify properties and REITs that have not been invested in yet
        properties_without_investments = self.property_data[~self.property_data['property_id'].isin(self.investment_data['property_id'])]
        reits_without_investments = self.reit_data[~self.reit_data['reit_id'].isin(self.investment_data['reit_id'])]

        return {'users_without_investments': users_without_investments, 'properties_without_investments': properties_without_investments, 'reits_without_investments': reits_without_investments}

    def recommend_improvements(self):
        areas_for_improvement = self.identify_areas_for_improvement()

        # Generate recommendations for users who have not made an investment yet
        for user_id in areas_for_improvement['users_without_investments']['user_id']:
            generate_recommendations(user_id)

        # Automate management for properties and REITs that have not been invested in yet
        for property_id in areas_for_improvement['properties_without_investments']['property_id']:
            automate_management(property_id)
        for reit_id in areas_for_improvement['reits_without_investments']['reit_id']:
            automate_management(reit_id)
```