tests/test_data_analytics.py
```python
import unittest
from src.data_analytics import DataAnalytics
from src.user import User
from src.investment import Investment
from src.property import Property
from src.reit import REIT

class TestDataAnalytics(unittest.TestCase):

    def setUp(self):
        self.data_analytics = DataAnalytics()
        self.user = User(user_id=1, name="John Doe", investment_preferences="Commercial", risk_tolerance="High")
        self.investment = Investment(investment_id=1, user_id=1, property_id=1, reit_id=1, investment_amount=100000, investment_date="2022-01-01")
        self.property = Property(property_id=1, property_type="Commercial", property_value=200000, rental_income=2000)
        self.reit = REIT(reit_id=1, reit_name="Top REIT", reit_value=500000)

    def test_analyze_data(self):
        result = self.data_analytics.analyze_data(self.user.user_id)
        self.assertIsNotNone(result)

    def test_generate_recommendations(self):
        result = self.data_analytics.generate_recommendations(self.user.user_id)
        self.assertIsNotNone(result)

    def test_automate_management(self):
        result = self.data_analytics.automate_management(self.investment.investment_id)
        self.assertIsNotNone(result)

if __name__ == '__main__':
    unittest.main()
```